headers
git-svn-id: https://osirix.svn.sourceforge.net/svnroot/osirix@3380 d300eccd-f00d-0410-ae38-fa88e4616176

+++ plugins/Invert - WindowController/InvertFilter.m
@@ -83,7 +83,7 @@
 		[viewerController endWaitWindow: waitWindow];
 		
 		// Update the current displayed WL & WW : we just inverted the image -> invert the WL !
-		long wl, ww;
+		float wl, ww;
 		
 		[[viewerController imageView] getWLWW: &wl :&ww];
 		if( [curPix isRGB]) wl = 255-wl;
